Added `GalleryItem`, overhauled `Gallery`
